Changed atar type from int to float

+++ sql/createTables.sql
@@ -33,7 +33,7 @@ CREATE TABLE courses (
 CREATE TABLE programs(
     program_code VARCHAR(20) PRIMARY KEY,
     name VARCHAR(200) NOT NULL,
-    atar SMALLINT CHECK(atar >= 0 AND atar <= 100),
+    atar FLOAT CHECK(atar >= 0 AND atar <= 100),
     career VARCHAR(20),
     duration SMALLINT CHECK(duration > 0 AND duration < 10),
     mod VARCHAR(100)
